SUPER IMPORTANT COMMIT: finally discover the DX & GL ASSIMP different flags! This will allow literally read any/assimp 3D animated format.

+++ CoreEngine_Tutorials/DXENGINE_021/core/src/dxWinSystemClass.cpp
@@ -166,7 +166,7 @@ int dxWinSystemClass::APPLICATION_MAIN_LOOP()		// [RUN] - MAIN "INFINITE" LOOP!
     }
 
     WOMA_LOGManager_DebugMSG("msg.wParam: %d\n", msg.wParam);
-	return msg.wParam; //return the PostQuitMessage (message code)
+	return (int)msg.wParam; //return the PostQuitMessage (message code)
 }
 
 //----------------------------------------------------------------------------
